+++ src/util/templates/ChainedTemplateEngine.ts
@@ -1,4 +1,5 @@
-import type { Template, TemplateEngine } from './TemplateEngine';
+import type { TemplateEngineInput } from './TemplateEngine';
+import { TemplateEngine } from './TemplateEngine';
 import Dict = NodeJS.Dict;
 
 /**
@@ -9,7 +10,7 @@ import Dict = NodeJS.Dict;
  * All subsequent engines will be called with no template parameter.
  * Contents will still be passed along and another entry will be added for the body of the previous output.
  */
-export class ChainedTemplateEngine<T extends Dict<any> = Dict<any>> implements TemplateEngine<T> {
+export class ChainedTemplateEngine<T extends Dict<any> = Dict<any>> extends TemplateEngine<T> {
   private readonly firstEngine: TemplateEngine<T>;
   private readonly chainedEngines: TemplateEngine[];
   private readonly renderedName: string;
@@ -19,6 +20,7 @@ export class ChainedTemplateEngine<T extends Dict<any> = Dict<any>> implements T
    * @param renderedName - The name of the key used to pass the body of one engine to the next.
    */
   public constructor(engines: TemplateEngine[], renderedName = 'body') {
+    super();
     if (engines.length === 0) {
       throw new Error('At least 1 engine needs to be provided.');
     }
@@ -27,12 +29,14 @@ export class ChainedTemplateEngine<T extends Dict<any> = Dict<any>> implements T
     this.renderedName = renderedName;
   }
 
-  public async render(contents: T): Promise<string>;
-  public async render<TCustom = T>(contents: TCustom, template: Template): Promise<string>;
-  public async render<TCustom = T>(contents: TCustom, template?: Template): Promise<string> {
-    let body = await this.firstEngine.render(contents, template!);
+  public async canHandle(input: TemplateEngineInput<T>): Promise<void> {
+    return this.firstEngine.canHandle(input);
+  }
+
+  public async handle({ contents, template }: TemplateEngineInput<T>): Promise<string> {
+    let body = await this.firstEngine.handle({ contents, template });
     for (const engine of this.chainedEngines) {
-      body = await engine.render({ ...contents, [this.renderedName]: body });
+      body = await engine.handleSafe({ contents: { ...contents, [this.renderedName]: body }});
     }
     return body;
   }